use config_file path, add args and uninstall method
